Put back test in jedi ebuild

+++ dev-python/jedi/jedi-0.11.1.ebuild
@@ -15,16 +15,27 @@ SRC_URI="https://github.com/davidhalter/jedi/archive/v${PV}.tar.gz -> ${P}.tar.g
 LICENSE="MIT"
 SLOT="0"
 KEYWORDS="amd64 x86"
-IUSE="doc"
+IUSE="doc test"
+REQUIRED_USE="${PYTHON_REQUIRED_USE} test? ( || ( python_targets_python3_4 python_targets_python3_5 ) !python_targets_python2_7 )"
+
+RESTRICT="test"
 
 DEPEND="
 	>=dev-python/parso-0.1.1
 	dev-python/setuptools[${PYTHON_USEDEP}]
-	doc? ( dev-python/sphinx )"
+	doc? ( dev-python/sphinx )
+	test? ( dev-python/pytest[${PYTHON_USEDEP}] )"
 
 python_prepare_all() {
-	rm -rf test #the tests only support python 3.5 and above
+	if ! use test ; then
+		rm -rf test #the tests only support python 3.5 and above
+	fi
 	distutils-r1_python_prepare_all
+}
+
+python_test() {
+	PYTHONPATH="${PYTHONPATH%:}${PYTHONPATH+:}${S}/test" py.test -v -v test \
+		|| die "Tests failed under ${EPYTHON}"
 }
 
 python_compile_all() {
